requirements.txt
ultralytics
psycopg2-binary
requests
python-dotenv